Add flag to compress chunks in hdf5 intermediate data
Note that `--compress-intermediate` has been configured to imply
`--chunk-intermediate`. This is because if compression is applied
without also triggering the code to specify the chunk shape to be 1
chunk per projection/sinogram, then `h5py` will guess the chunk shape to
write. This is often not as optimal as choosing 1 chunk per
projection/sinogram, in terms of compression ratio.