Switch from `logging` to `loguru`
Fixes #288

Note that some tests require certain information to be written to the
concise logfile. Some of this information may not be appropriate for the
concise logfile, and will potentially be moved to the verbose logfile.
However, for now, the information required to make the tests pass has
been put into the concise logfile.