Write block progress bar to concise logs
It was tricky to get the `tqdm` progress bar to be written to the
concise logfile only once for each percentage update, and to be shown
only for the rank 0 process in the terminal, while also not duplicating
the block loop multiple times to handle the case of rank 0 vs. other
ranks.

It was found to be possible to get a string representing the `tqdm`
progress bar. In order to not display the same progress bar percentage
multiple times in the terminal (or write the same progress bar
percentage multiple times to the concise logfile), this string is passed
to the logger object, which has configuration to handle log messages
from multiple MPI processes.

Note that the idea of redirecting the output of `tqdm` to `/dev/null`
and manually handling the progress bar string came from
https://stackoverflow.com/a/75059930.