Change chunk intermediate flag to specify frames per-chunk
Setting the `--frames-per-chunk` flag to 0 will write the intermediate
data as contiguous, and any value > 0 will be the number of frames
(projections or sinograms) per-chunk.

Thus, this preserves the ability to turn chunking on/off when writing
intermediate data, but also adds the ability to choose the number of
frames per-chunk.