Use 1 frame per chunk if flag value exceeds slicing dim length
If the value provided for the `--frames-per-chunk` flag is larger than
the length of the data's slicing dimension, then the number of frames
per-chunk will be set to 1 (and will be logged in the verbose logfile).

This is to avoid `h5py` errors with trying to create a chunk whose shape
exceeds the boundaries of the data being written.