Add block interfaces and base block class
The motivation behind this change is that it would be useful to be able
to use method wrapper objects (implementors of `MethodWrapper`) to
execute methods in both:
- the high throughput runner `TaskRunner`
- the parameter sweep runner `ParamSweepRunner`

The `DataSetBlock` class contains code that is mostly compatible with a
parameter sweep run. However, one incompatibility is the
`DataSetBlock.data.setter()` method, which places a constraint on the
data involving the slicing dimension of the data. In parameter sweep
runs there is only a single process being run, and so the concept of
slicing dimension isn't relevant.

Given that `DataSetBlock` contains a fair amount of logic that *is*
usable in a parameter sweep run, after various iterations of testing out
what could work well, the logic in `DataSetBlock` that is usable across
both high throughput runs and parameter sweep runs has been extracted
into a separate `BaseBlock` class.

A high-level overview of the organisation of block-related functionality
now is:
- the various functionalites needed for a block type to be processable
  by implementors of `MethodWrapper` have been organised into three
  separate protocols: `BlockData`, `BlockTransfer`, and `BlockIndexing`
- `BaseBlock` provides "typical implementations" for `BlockData` and
  `BlockTransfer`, which are reused by `DataSetBlock` via inheritance
- thus, `BaseBlock` contains the code for "typical" implementations of
  methods that are common to blocks across both high throughput runs and
  parmeter sweep runs
- `DataSetBlock` contains only the code specific to processing blocks in
  high throughput runs
- in the future, the "typical implementations" contained in `BaseBlock`
  can be reused by a new block type that will be used in parameter sweep
  runs