Infer shape of array holding all sweep results from first block shape + no of sweep
Before, the array to hold all sweep results was inferred from the
`single_shape` value that was given when the writer object was first
created. This had the implicit assumption that the shape of a single
sweep result would be known at the time the writer object is created.

This assumption holds when the method being executed in the sweep
doesn't change the shape of the data during processing (most methods
adhere to this). However, this assumption doesn't hold when the method
executed in the sweep does change the shape of the data during
processing (for example, reconstruction methods).

Therefore, the writer now infers the shape of a single sweep result when
the first block to be written has been given. This can then can be used
to infer the shape of the array to hold all sweep results for both cases
(when the method being executed in the sweep does and doesn't change the
shape of the data).