Define comm objects in CLI
Where possible, when a communicator object is used, it should be by
passing one of these two communicator objects that are created in the
CLI:
- global communicator
- communicator for method wrapper objects to use during method execution
  (this may or may not be the global communicator in the future, based
  on what type of run is being executed)

Only where a communicator object is being used and it must be a specific
communicator, should a communicator object not be passed into the
function.

For example, the calculation of the GPU ID associated with a process
must be based on the global and "split" communicators, so the
`get_gpu_id()` function doesn't provide a communicator parameter to be
passed in.