Set `offset` param of images wrapper in sweep runner
When running the `save_to_images` method in the context of httomo, the
`offset` parameter doesn't make sense to be exposed to the user through
the pipeline configuration. This is because for high throughput runs,
the `offset` is always inferred by the global index of the block being
saved.

Setting the `offset` parameter of the image saver wrapper from the sweep
runner enabled the correct saving of the output of the different
processes in a parameter sweep run, but this required a small
modification to the image saver wrapper. The modification is that, now,
the image saver wrapper will infer the `offset` from the block's global
index only when the `offset` isn't provided.

So, for all pipeline cases (high throughput and sweep), the `offset`
parameter is omitted from the pipeline file. In the case of high
throughput runs, the `offset` will be calculated as it was before via a
block's global index. In the case of a parameter sweep run, the `offset`
is explicitly set in the sweep runner for each process.