Account for two copies of chunk needed when deciding store backing
When a reader is created from a writer, a padded version of the chunk is
held simultaneously alongside the unpadded version of the chunk. This
means that the memory requirements for the store to be backed by RAM are
at least the size of two copies of a chunk.

This change modifies the writer to account for the need of approximately
two copies of the chunk, when it is deciding if the store should be
backed by RAM or an hdf5 file.

Note that the padded chunk shape should be the padded chunk shape needed
for the next section, as opposed to the padded chunk shape for the
current section, which is what is currently in place. So the writer will
need to be tweaked again at some point to account for that difference.