Account for deallocated padded float32 array in memory estimation
The padded `float32` array is cast to `complex64`, and the variable
originally assigned to the `float32` version is reassigned to the
`complex64` version.

The reassigned variable was the only thing referring to the `float32`
array, so the ref count of it drops to 0 and is deallocated.

The padded `float32` array gets deallocated early enough in the method
that it doesn't contribute to the *peak* memory usage of the paganin
filter method, which was verified by checking:
- the output of `LineProfileHook` when running the method
- the allocations and deallocations reported by the custom
  `MaxMemoryHook` in the tests during the method's execution