Modify wrapper factory function to pass preview to distortion correction wrapper class
The bulk of the changes were simply to provide the preview config
parameter to tests that used the wrapper factory function.

A non-trivial change was to modify the transform layer's insertion of
specific methods to not use the wrapper factory function, and instead
create the wrapper objects directly using the appropriate wrapper
classes.

The original motivation for this was that the preview config information
isn't available to the transform layer object, so it wouldn't be simple
to get that information to pass to the wrapper factory function.

Another line of reasoning for this change was that, due to the transform
layer knowing exactly which methods it wants to insert, it seemed
unnecessary to create the wrapper objects via the factory function: it
knows which specific methods to insert, and thus the wrapper object it
needs to create for each of the methods to insert is already known at
that point in time.