Fix incorrect types relating to darks/flats config
After doing the direct translation from what was done before into the
parsing being done now to handle the `auto` value, some type-hints were
found to be not quite correct.

If the `darks` or `flats` key is specified in the loader config (whose
value would be a dict), then both keys `file` and `data_path` need to be
in the dict. Hence, the `NotRequired[str]` type-hint was incorrect.

On the other hand, what is possible is that the `darks` or `flats` key
can be omitted entirely. In which case, there is no dict for configuring
darks/flats, and it has a `None` value instead. Hence,
`Optional[DarksFlatsParam]` is appropriate.

Finally, there shouldn't be an `image_key_path` key inside the dict
config for darks/flats. Hence, the use of `get("image_key_path",
image_key_path)` was unnecessary, because the key would never exist and
thus the get operation would always return the default value.