Replace CLI no-args test to check correct behaviour
Prior to this change the test was checking that, when running the CLI
with no args, if an error was raised, to ensure a specific exit code was
returned.

However, when the CLI is given no args, an error is not expected to be
raised; instead, the help message is expected to be displayed.
Therefore, the test was checking for behaviour that the CLI never
exhibited.

Also, this change replaces the subprocess-launching in the test with
testing functionality provided by `click` to run the CLI.