Add MPI abort handler for unhandled exceptions
This is a catch-all approach intended to make sure that, if any MPI
processes are killed due to hitting an exception that was not explicitly
handled by httomo, then all MPI processes are killed. This is to catch
any cases where one or more MPI processes get killed but not all of
them, resulting in a deadlock where some processes are alive and waiting
for other dead processes (and thus are waiting indefinitely).

If such a case is encountered where there's somehere in the httomo
source code that, due to a lack of explicitly handling an exception, a
deadlock can occur, it's advisable to make appropriate changes to handle
the exception at the specific part in the code where the exception can
occur. This catch-all approach is intended as a way to prefer
terminating program execution over a deadlock, to then help with
narrowing down the root cause, it's not intended as a replacement for
more finely-grained exception handling.