Add test to check min block len communication based on MPI reduce call
The reason that checking if the reduce method on the communicator was
called is used as a proxy for checking if the minimum block length was
communicated is because the former can be done with one MPI process (via
mocking the size of the communicator). On the other hand, checking
directly if the minimum block length was communicated or not would
require running two or more processes, which is less convenient in unit
tests.