Create output directory arg if it doesn't exist
As described by the [`click`
documentation](https://click.palletsprojects.com/en/stable/api/#click.Path),
in the case where `exists=False` and the directory doesn't exist, checks
such as `writable` are silently skipped.

The added try/except is for this case where the output directory doesn't
exist and is attempted to be created but the user has insufficient
permissions to do so, where `click` won't perform the error handling for
us.