Don't account for padding in reader's chunk shape and global index
Fixes #643

The `global_index` on `DataSetStoreReader` has also been changed to not
account for padding, due to it being somewhat connected to the padding,
and being able to find a way to fix things by including it in the set of
possible things to modify.

While the loader does account for padding in its `global_index`
(meaning, the `DataSetStoreReader` isn't exposing shapes and indices
regarding padding in exactly the same manner as the loader, even with
these changes), the block splitter doesn't rely on `global_index` on the
implementor of `DataSetSource` that it's given, so this shouldn't cause
any inconsistencies between how the block splitter uses either a loader
or a dataset store reader as a data source.

However, whether `global_index` should or shouldn't account for padding
indeed may be something to revisit if other issues arise with the
dataset store relating to padding.

Note that, while this change prevents the error in the linked issue from
occurring, the padding applied to blocks when the dataset store is
backed by either RAM or an hdf5 file is now incorrect.