Remove explicitly setting frames per chunk to 1
The frames per chunk need not be set to 1 unless the minimum block
length is 1. So, the `if` statement that remains, which sets the frames
per chunk to be the same as the minimum block length under the
appropriate condition, is sufficient to ensure a frames per chunk value
that will work across all blocks within a section.